requirements.txt
anyjson==0.3.3
argon2-cffi==18.3.0
arrow==0.12.1
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
boto3==1.9.35
botocore==1.12.35
certifi==2018.10.15
cffi==1.11.5
colorama==0.3.9
colour==0.1.5
cryptography==2.3.1
decorator==4.1.2
docutils==0.14
furl==1.2.1
idna==2.7
infinity==1.4
inflection==0.3.1
intervals==0.8.1
intrinio-sdk==4.0.0
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jmespath==0.9.3
joblib==0.12.5
more-itertools==3.2.0
numpy==1.13.1
orderedmultidict==1.0
pandas==0.20.3
passlib==1.7.1
path.py==10.3.1
phonenumbers==8.9.11
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
Quandl==3.1.0
requests==2.14.2
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.13
SQLAlchemy-Utils==0.33.3
tqdm==4.15.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
win-inet-pton==1.0.1
wincertstore==0.2
